requirements.txt
# POST TOOL - Python Dependencies

# AI Integration
anthropic>=0.7.0

# Git Operations  
GitPython>=3.1.40

# File System Monitoring
watchdog>=3.0.0

# Markdown Processing
python-frontmatter>=1.1.0

# HTTP Requests
requests>=2.31.0

# Environment Configuration
python-dotenv>=1.0.0

# Date/Time Utilities
python-dateutil>=2.8.2

# JSON Schema Validation
jsonschema>=4.17.0